\name{CStruct-class}
\Rdversion{1.1}
\docType{class}
\alias{CStruct-class}

\title{Class "CStruct"}
\description{
This is a virtual class that is used as a root or top-level class to
identify that sub-class arise from copies of C or C++ structures.
}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "CStruct" in the signature.
}
\references{
}
\author{
Duncan Temple Lang
}

\seealso{

}
\examples{

}
\keyword{classes}
