\name{EnumValue-class}
\docType{class}
\alias{EnumValue-class}
\title{Class "EnumValue" representing a enumerated constant}
\description{This class represents the definition of an enumerated
  constant in C.}
%\section{Objects from the Class}{}
\section{Slots}{
	 \describe{
    \item{\code{.Data}:}{Object of class \code{"integer"}}
  }
}
\section{Extends}{
Class \code{"\linkS4class{SymbolicConstant}"}, directly.
Class \code{"\linkS4class{EnumerationValue}"}, directly.
Class \code{"\linkS4class{integer}"}, by class "SymbolicConstant", distance 2.
Class \code{"\linkS4class{vector}"}, by class "SymbolicConstant", distance 3.
Class \code{"\linkS4class{numeric}"}, by class "SymbolicConstant", distance 3.
Class \code{"\linkS4class{data.frameRowLabels}"}, by class "SymbolicConstant", distance 3.
}
\section{Methods}{
No methods defined with class "EnumValue" in the signature.
}
\references{}
\author{Duncan Temple Lang}
\seealso{
}
\examples{
showClass("EnumValue")
}
\keyword{classes}
